Refactor getSecretSource to secret-reader file
