+++ test/k6/src/api/token-generator.js
@@ -4,7 +4,7 @@ import encoding from "k6/encoding";
 import * as config from "../config.js";
 import { stopIterationOnFail } from "../errorhandler.js";
 import * as apiHelpers from "../apiHelpers.js";
-import { getFromSecretSource } from "../setup.js";
+import { getFromSecretSource } from "../secret-reader.js";
 
 const environment = (__ENV.altinn_env || '').toLowerCase(); // Fallback value for when k6 inspect is run in script validation (env var evaluation yields 'undefined' in this phase)
